Important modification of the interface in order to soimplify the add of new methods or commands to the interface For matlab and python interface there is now a unique source which are the gf_*.cc files. All the documentations and command files (mfiles and python file) are produced automatically.
This should be extended to the Scilab interface ...

Please check that matlab and python are still perfectly working. I probably make some mistakes ...

The official documentation is now the one in doc/sphinx. It should be uploaded soon to be accessible by the web. The one for the Matlab and Python interfaces are automatic. They should probably be improved.