Typo and UTF-8
It look like there is a typo in the conditional include, I guessing you
want to include only if not defined, but there is not the same number of
underscore in define.

Also one of the files are in ISO-8859 (latin1) I guess, which make clang
show warnings for `é` and a few other character, do not change content
but change to utf8, it seems to fix the warning. This has the side
effect of making the diff show the correct characters instead of
mojibake.